Restructure the project to use a single entrypoint script managed by Cobra.
