+++ aoc/cmd/root.go
@@ -1,0 +1,104 @@
+package cmd
+
+import (
+	"aoc/pkg/advent"
+	"aoc/pkg/year_2022"
+	"github.com/spf13/cobra"
+	"log"
+)
+
+var rootCmd = &cobra.Command{
+	Use:   "my-aoc",
+	Short: "Compute and submit Advent Of Code solutions",
+	Long:  "my-aoc lets you submit the solution to any day of Advent Of Code by running your text input against the solutions provided in this project.",
+	Run:   Run,
+}
+
+var Submit bool = false
+
+func init() {
+	rootCmd.PersistentFlags().IntP("day", "d", 1, "The day to run solutions for.")
+	rootCmd.PersistentFlags().IntP("year", "y", 2022, "The year to run solutions for.")
+	rootCmd.PersistentFlags().IntP("part", "p", 1, "The part to run solution for.")
+	rootCmd.PersistentFlags().BoolVarP(&Submit, "submit", "s", false, "should submit")
+
+	_ = rootCmd.MarkFlagRequired("verbose")
+	_ = rootCmd.MarkFlagRequired("day")
+	_ = rootCmd.MarkFlagRequired("year")
+	_ = rootCmd.MarkFlagRequired("part")
+}
+
+func Run(cmd *cobra.Command, args []string) {
+	flags := cmd.Flags()
+	day, _ := flags.GetInt("day")
+	year, _ := flags.GetInt("year")
+	part, _ := flags.GetInt("part")
+
+	if year != 2022 {
+		log.Fatal("Currently only have solutions for 2022. Please use 2022 as year flag or skip that flag altogether.")
+	}
+
+	inputs, err := advent.GetInputs(day, year)
+	if err != nil {
+		log.Fatalf("Could not fetch inputs: %s", err.Error())
+	}
+
+	switch day {
+	case 1:
+		{
+			year_2022.Solve(day, year, part, inputs, &year_2022.Day01{}, Submit)
+			break
+		}
+	case 2:
+		{
+			year_2022.Solve(day, year, part, inputs, &year_2022.Day02{}, Submit)
+			break
+		}
+	case 3:
+		{
+			year_2022.Solve(day, year, part, inputs, &year_2022.Day03{}, Submit)
+			break
+		}
+	case 4:
+		{
+			year_2022.Solve(day, year, part, inputs, &year_2022.Day04{}, Submit)
+			break
+		}
+	case 5:
+		{
+			year_2022.Solve(day, year, part, inputs, &year_2022.Day05{}, Submit)
+			break
+		}
+		//case 6:
+		//	{
+		//		year_2022.Solve(day, year, part, inputs, &year_2022.Day06{}, Submit)
+		//		break
+		//	}
+		//case 7:
+		//	{
+		//		year_2022.Solve(day, year, part, inputs, &year_2022.Day07{}, Submit)
+		//		break
+		//	}
+		//case 8:
+		//	{
+		//		year_2022.Solve(day, year, part, inputs, &year_2022.Day08{}, Submit)
+		//		break
+		//	}
+		//case 9:
+		//	{
+		//		year_2022.Solve(day, year, part, inputs, &year_2022.Day09{}, Submit)
+		//		break
+		//	}
+		//case 10:
+		//	{
+		//		year_2022.Solve(day, year, part, inputs, &year_2022.Day10{}, Submit)
+		//		break
+		//	}
+	}
+}
+
+func Execute() {
+	if err := rootCmd.Execute(); err != nil {
+		log.Fatal(err)
+	}
+}
