+++ aoc/go.mod
@@ -2,10 +2,9 @@ module aoc
 
 go 1.18
 
+require github.com/spf13/cobra v1.6.1
+
 require (
-	github.com/joho/godotenv v1.4.0 // indirect
-	github.com/segmentio/fasthash v1.0.3 // indirect
-	github.com/yanatan16/itertools v0.0.0-20160513161737-afd1891e0c4f // indirect
-	github.com/zyedidia/generic v1.2.0 // indirect
-	golang.org/x/exp v0.0.0-20220218215828-6cf2b201936e // indirect
+	github.com/inconshreveable/mousetrap v1.0.1 // indirect
+	github.com/spf13/pflag v1.0.5 // indirect
 )
