+++ aoc/main.go
@@ -1,0 +1,7 @@
+package main
+
+import "aoc/cmd"
+
+func main() {
+	cmd.Execute()
+}
