+++ aoc/pkg/year_2022/day-05.go
@@ -1,0 +1,60 @@
+package year_2022
+
+import (
+	"aoc/pkg/structs"
+)
+
+type Day05 struct {
+}
+
+func (d *Day05) Part1(inputs []structs.Group) string {
+
+	//contents := inputs[0].Contents
+	//rounds := strings.Split(contents, "\n")
+	//var count int32 = 0
+	//
+	//for _, round := range rounds {
+	//	if round == "" {
+	//		break
+	//	}
+	//	sections := strings.Split(round, ",")
+	//	section1 := FromStr(sections[0])
+	//	section2 := FromStr(sections[1])
+	//
+	//	pair := WorkPair{
+	//		Section1: section1,
+	//		Section2: section2,
+	//	}
+	//	if pair.overlapsCompletely() {
+	//		count += 1
+	//	}
+	//}
+	//return fmt.Sprintf("%d", count)
+	return ""
+}
+
+func (d *Day05) Part2(inputs []structs.Group) string {
+
+	//contents := inputs[0].Contents
+	//rounds := strings.Split(contents, "\n")
+	//var count int32 = 0
+	//
+	//for _, round := range rounds {
+	//	if round == "" {
+	//		break
+	//	}
+	//	sections := strings.Split(round, ",")
+	//	section1 := FromStr(sections[0])
+	//	section2 := FromStr(sections[1])
+	//
+	//	pair := WorkPair{
+	//		Section1: section1,
+	//		Section2: section2,
+	//	}
+	//	if pair.overlaps() {
+	//		count += 1
+	//	}
+	//}
+	//return fmt.Sprintf("%d", count)
+	return ""
+}
